+++ xarray/core/combine.py
@@ -2,6 +2,7 @@ import itertools
 import warnings
 from collections import Counter
 
+import cftime
 import pandas as pd
 
 from . import dtypes
@@ -50,12 +51,18 @@ def _ensure_same_types(series, dim):
     if series.dtype == object:
         types = set(series.map(type))
         if len(types) > 1:
+            cftimes = any(issubclass(t, cftime.datetime) for t in types)
+
             types = ", ".join(t.__name__ for t in types)
-            raise TypeError(
+
+            error_msg = (
                 f"Cannot combine along dimension '{dim}' with mixed types."
                 f" Found: {types}."
-                f" Setting `use_cftime=True` may fix this issue."
             )
+            if cftimes:
+                error_msg = f"{error_msg} Setting `use_cftime=True` may fix this issue."
+
+            raise TypeError(error_msg)
 
 
 def _infer_concat_order_from_coords(datasets):